# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Maria Chat API
  version: "1.0.0"
  description: >
    API do backend do Maria Chat (atendimento jurídico da DPERJ).
    Grupos: webhook do WhatsApp, chat público, autenticação, painel admin
    (prefixo /admin, JWT), rotas internas do fluxo (/api/*, chamadas pelo engine)
    e mocks de teste (/mock/*). Rotas do painel exigem Bearer JWT; mutações
    exigem papel admin.
servers:
  - url: http://maria-chat-prod-alb-554264540.us-east-1.elb.amazonaws.com
    description: Produção (AWS ALB)
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Erro:
      type: object
      properties:
        erro: { type: string }
...
    CpfReq:
      type: object
      required: [cpf]
      properties:
        cpf: { type: string }
...
    AgendamentosConsultaResp:
      type: object
      properties:
        tem_agendamentos: { type: boolean }
        agendamentos:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              tipo: { type: string }
              data: { type: string, format: date-time }
              local: { type: string, nullable: true }
        lista: { type: string, description: "texto formatado, 1 linha por agendamento" }
    AgendamentoDetalheReq:
      type: object
      properties:
        agendamento_sel: { type: string, description: "id completo ou índice (1,2,...)" }
        agendamentos: { type: string, description: "JSON da consulta anterior (para mapear índice)" }
    AgendamentoDetalheResp:
      type: object
      required: [encontrado]
      properties:
        encontrado: { type: boolean }
        id: { type: string }
        tipo: { type: string }
        data: { type: string, format: date-time }
        local: { type: string, nullable: true }
        status: { type: string }
        assistido: { type: string }
        idAssunto:
          type: integer
          nullable: true
          description: "id numérico do assunto (Verde) — presente quando o agendamento veio do Verde; permite passar direto pra /api/assunto/documentos sem nova consulta ao Verde (issue maria-ia#20260234)."
    AgendamentoDetalheRicoReq:
      type: object
      required: [idEvento]
      properties:
        idEvento: { type: string }
    AgendamentoDetalheRicoResp:
      type: object
      required: [encontrado]
      description: "campos preservam os nomes originais do Verde (issue maria-ia#111, gateway#21)"
      properties:
        encontrado: { type: boolean }
        idEvento: { type: integer }
        idAssistido: { type: integer }
        idOrgao: { type: integer }
        idCaso: { type: integer, nullable: true }
        nomeAssistido: { type: string }
        nomeOrgao: { type: string }
        endereco: { type: string }
        data: { type: string }
        horario: { type: string }
        documentosNecessarios: { type: string, nullable: true }
        nomeAssunto: { type: string, nullable: true }
    AgendamentoVagasReq:
      type: object
      required: [idEvento]
      properties:
        idEvento: { type: string }
    AgendamentoVagasResp:
      type: object
      properties:
        tem_vagas: { type: boolean }
        vagas:
          type: array
          items:
            type: object
            properties:
              idIntervalo: { type: integer }
              data: { type: string }
              hora: { type: string }
        lista: { type: string, description: "texto formatado, 1 linha por vaga" }
    AgendamentoVagaDetalheReq:
      type: object
      properties:
        vaga_sel: { type: string, description: "índice (1,2,...) da lista de vagas anterior" }
        vagas: { type: string, description: "JSON da consulta de vagas anterior (resolve sem re-fetch)" }
    AgendamentoVagaDetalheResp:
      type: object
      required: [encontrada]
      properties:
        encontrada: { type: boolean }
        idIntervalo: { type: integer }
        data: { type: string }
        hora: { type: string }
    AgendamentoReagendarReq:
      type: object
      required: [idAgendamento, configuracaoIntervaloAgenda, dataNova]
      properties:
        idAgendamento: { type: string }
        configuracaoIntervaloAgenda: { type: string, description: "idIntervalo da vaga escolhida" }
        dataNova: { type: string }
        horaNova: { type: string, description: "combinada com dataNova antes de enviar ao Verde (\"DD/MM/YYYY HH:mm\")" }
    AgendamentoReagendarResp:
      type: object
      required: [sucesso]
      properties:
        sucesso: { type: boolean }
    AgendamentoDesmarcarReq:
      type: object
      required: [idAgendamento, idPessoa]
      properties:
        idAgendamento: { type: string }
        idPessoa: { type: string }
    AgendamentoDesmarcarResp:
      type: object
      required: [sucesso]
      properties:
        sucesso: { type: boolean }
        emailEnviado: { type: boolean, description: "204 do Verde (true) vs 200 sucesso parcial (false, e-mail falhou)" }
    VerificarDuplicadosReq:
      type: object
      required: [idPessoa, idAssunto]
      properties:
        idPessoa: { type: integer }
        idAssunto: { type: integer }
    VerificarDuplicadosResp:
      type: object
      required: [tem_duplicado]
      properties:
        tem_duplicado: { type: boolean }
        idAgendamentoDuplicado: { type: integer, nullable: true }
        dataDuplicado: { type: string, nullable: true }
    AgendarReq:
      type: object
      required: [idPessoa, idOrgao, idAssunto, idIntervalo, dataVaga]
      properties:
        idPessoa: { type: integer }
        idOrgao: { type: integer }
        idAssunto: { type: integer }
        idIntervalo: { type: integer, description: "vindo de orgao/primeiro-atendimento (primeira_vaga_intervalo) ou agendamentos/vagas" }
        dataVaga: { type: string, description: "\"DD/MM/YYYY HH:mm\"" }
        resumoAtendimento: { type: string }
        textoComplemento: { type: string }
    AgendarResp:
      type: object
      required: [sucesso]
      properties:
        sucesso: { type: boolean }
        idAgendamento: { type: integer, nullable: true }
...
security:
  - bearerAuth: []
paths:
...
  /api/agendamentos/consultar:
    post:
      tags: [Interno (fluxo)]
      summary: Agendamentos em aberto do assistido (tabela Agendamento)
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/CpfReq' } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentosConsultaResp' } } } }
        "503": { description: banco não configurado, content: { application/json: { schema: { $ref: '#/components/schemas/Erro' } } } }
  /api/agendamentos/detalhe:
    post:
      tags: [Interno (fluxo)]
      summary: Detalhe de 1 agendamento (por id ou índice da lista anterior)
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoDetalheReq' } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoDetalheResp' } } } }
        "503": { description: banco não configurado, content: { application/json: { schema: { $ref: '#/components/schemas/Erro' } } } }
  /api/agendamentos/detalhe-rico:
    post:
      tags: [Interno (fluxo)]
      summary: "Detalhe rico de 1 agendamento (endereço, documentos, órgão, idAssistido) — proxy GET api/agendamento/{idEvento} do gateway (issue #111)"
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoDetalheRicoReq' } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoDetalheRicoResp' } } } }
        "503": { description: banco não configurado, content: { application/json: { schema: { $ref: '#/components/schemas/Erro' } } } }
  /api/agendamentos/vagas:
    post:
      tags: [Interno (fluxo)]
      summary: "Vagas disponíveis pra reagendar — proxy GET api/agendamento/vagas/{idEvento} do gateway"
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoVagasReq' } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoVagasResp' } } } }
        "503": { description: banco não configurado, content: { application/json: { schema: { $ref: '#/components/schemas/Erro' } } } }
  /api/agendamentos/vaga-detalhe:
    post:
      tags: [Interno (fluxo)]
      summary: Resolve a vaga escolhida do JSON que /vagas já retornou (índice, zero re-fetch)
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoVagaDetalheReq' } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoVagaDetalheResp' } } } }
        "503": { description: banco não configurado, content: { application/json: { schema: { $ref: '#/components/schemas/Erro' } } } }
  /api/agendamentos/reagendar:
    post:
      tags: [Interno (fluxo)]
      summary: "Reagenda um agendamento — proxy POST api/agendamento/reagendar do gateway"
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoReagendarReq' } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoReagendarResp' } } } }
        "503": { description: banco não configurado, content: { application/json: { schema: { $ref: '#/components/schemas/Erro' } } } }
  /api/agendamentos/desmarcar:
    post:
      tags: [Interno (fluxo)]
      summary: "Desmarca um agendamento — proxy POST api/agendamento/desmarcar do gateway (ação irreversível, dispara e-mail real)"
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoDesmarcarReq' } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/AgendamentoDesmarcarResp' } } } }
        "503": { description: banco não configurado, content: { application/json: { schema: { $ref: '#/components/schemas/Erro' } } } }
  /api/agendamentos/verificar-duplicados:
    post:
      tags: [Interno (fluxo)]
      summary: "Checa agendamento duplicado do mesmo assunto — proxy GET api/agendamento/verificar-duplicados do gateway"
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/VerificarDuplicadosReq' } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/VerificarDuplicadosResp' } } } }
  /api/agendamentos/agendar:
    post:
      tags: [Interno (fluxo)]
      summary: "Cria agendamento novo (primeiro atendimento, com hora marcada) — proxy POST api/agendamento/agendar do gateway"
      security: []
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/AgendarReq' } } } }
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/AgendarResp' } } } }
...